[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
<add g3 doc link>

Bug: 304529413
Test: N/A
Ignore-AOSP-First: The migration was already run for aosp and these files don't exist in aosp - https://android-review.git.corp.google.com/c/platform/packages/modules/common/+/2965104
Change-Id: I7713999ec2238b136c05bf24bb496ef428752e91