Define aosp_mainline_modules(placeholder)
We're trying to build Android system image which is defined in
Android.bp. It's cumbersome to set visibility for every apex module. So
define an aggregated module for mainline modules.

It's a placeholder to unblock automerge

Bug: 321000103
Test: m nothing
Ignore-AOSP-First: new type of visibility doesn't exist in udc
Change-Id: I3fb278feaa015ecbe3bb3a38440561533af8ad46
Merged-In: I3fb278feaa015ecbe3bb3a38440561533af8ad46